requirements.txt
scrapy==2.11.2
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
requests==2.31.0
python-dotenv==1.0.1
lxml==5.2.1
cssselect==1.2.0
